requirements.txt
annotated-types==0.7.0
anyio==4.5.2
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2025.1.31
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
Django>=4.2,<5.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
executing==2.0.1
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.4.0
inflection==0.5.1
ipython==8.12.3
ipywidgets==8.1.3
jedi==0.19.1
joblib==1.4.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
numpy==1.24.4
ollama==0.4.7
packaging==24.1
pandas==2.0.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.1
threadpoolctl==3.5.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
wcwidth==0.2.13
widgetsnbextension==4.0.11
zipp==3.19.2